+++ xbird-cloud-netflix/xbird-cloud-netflix-zuul/src/main/java/com/xbird/cloud/netflix/zuul/ZuulApplication.java
@@ -1,0 +1,20 @@
+package com.xbird.cloud.netflix.zuul;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.netflix.hystrix.EnableHystrix;
+import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
+
+@SpringBootApplication
+@EnableEurekaClient
+@EnableZuulProxy
+@EnableHystrix
+@EnableHystrixDashboard
+public class ZuulApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(ZuulApplication.class, args);
+  }
+}