+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - [新增] 添加 Spring Cloud Netflix Hystrix 项目。
 - [新增] 添加 Spring Cloud Netfilx Turbine 项目。
 - [新增] 添加 Spring Cloud Netfilx Ribbon 项目。
+- [新增] 添加 Spring Cloud Netfilx Zuul 项目。